+++ src/evaluate/utils.rs
@@ -6,12 +6,14 @@ use std::{
 
 use rust_xlsxwriter::{Format, Worksheet};
 
+use crate::utils::FloatType;
+
 fn get_buf_reader(file_name: String) -> BufReader<File> {
     let file = File::open(file_name).expect("Failed to open file");
     BufReader::new(file)
 }
 
-pub fn get_input_data(test_number: u32) -> (usize, Vec<f64>) {
+pub fn get_input_data(test_number: u32) -> (usize, Vec<FloatType>) {
     let file_name = format!("./input/input{:03}.txt", test_number);
     let reader = get_buf_reader(file_name);
 
@@ -19,9 +21,9 @@ pub fn get_input_data(test_number: u32) -> (usize, Vec<f64>) {
     let first_line = lines.next().expect("Empty file").unwrap();
     let n: usize = first_line.trim().parse().expect("Invalid number");
 
-    let mut radiuses: Vec<f64> = Vec::with_capacity(n);
+    let mut radiuses: Vec<FloatType> = Vec::with_capacity(n);
     for line in lines {
-        let radius: f64 = line
+        let radius: FloatType = line
             .expect("Failed to read line")
             .trim()
             .parse()
@@ -33,7 +35,7 @@ pub fn get_input_data(test_number: u32) -> (usize, Vec<f64>) {
     return (n, radiuses);
 }
 
-pub fn get_jury_answer(test_number: u32) -> f64 {
+pub fn get_jury_answer(test_number: u32) -> FloatType {
     let file_name = format!("./output/out{:03}.txt", test_number);
     let reader = get_buf_reader(file_name);
 
@@ -44,7 +46,7 @@ pub fn get_jury_answer(test_number: u32) -> f64 {
         .unwrap()
         .trim_end()
         .to_string()
-        .parse::<f64>()
+        .parse::<FloatType>()
         .expect("Valid float jury answer.")
 }
 
@@ -52,10 +54,10 @@ pub fn write_row_block(
     worksheet: &Arc<Mutex<&mut Worksheet>>,
     row: u32,
     col: u16,
-    main_circle_radius: f64,
+    main_circle_radius: FloatType,
     is_valid: bool,
-    points: f64,
-    time: f64,
+    points: FloatType,
+    time: FloatType,
     format: &Format,
 ) {
     worksheet
@@ -80,6 +82,6 @@ pub fn write_row_block(
         .ok();
 }
 
-pub fn calculate_points(answer: f64, jury_answer: f64) -> f64 {
+pub fn calculate_points(answer: FloatType, jury_answer: FloatType) -> FloatType {
     ((2.0 - (answer / jury_answer)) * 100.0).max(0.0)
 }
