+++ src/ralgo/ralgo_result.rs
@@ -1,10 +1,10 @@
-use crate::circle::Circle;
+use crate::{circle::Circle, utils::FloatType};
 
 pub struct RalgoResult {
     pub ralgo_calls: u32,
     pub iterations: u32,
     pub calcfg_calls: u32,
-    pub main_circle_radius: f64,
+    pub main_circle_radius: FloatType,
     pub circles: Vec<Circle>,
 }
 
@@ -13,7 +13,7 @@ impl RalgoResult {
         ralgo_calls: u32,
         iterations: u32,
         calcfg_calls: u32,
-        main_circle_radius: f64,
+        main_circle_radius: FloatType,
         circles: Vec<Circle>,
     ) -> Self {
         RalgoResult {
